+++ client/src/darwingGrids.cpp
@@ -9,8 +9,9 @@
 #include "computationalLib/math_core/Dipoles.h"
 
 #include "../include/plotingUtils.h"
-#include "iolib/Printers.h"
+#include "common/Printers.h"
 
+using namespace printUtils;
 //template <class Tr>
 std::array<std::vector<double>, 2> squareGrid(double l, double r, double a = 0, double b = 0, bool center = false) {
     //long double lim= ceil(r+r/l);
