+++ utilLib/include/common/Printers.h
@@ -10,8 +10,13 @@
 #include <json/json.h>
 #include <valarray>
 #include <concepts>
+#include "common/typeCasts.h"
+#include "common/printUtils.h"
 /// printUtils namespace
 namespace printUtils {
+
+    using commonTypes::FloatType;
+    using printUtils::IosStatePreserve, printUtils::IosStateScientific;
     template<typename Iterator>
     //todo minimal req is bidirectional iterator
     //todo move near printers in utility(include only json)
@@ -26,24 +31,82 @@ namespace printUtils {
         return res;
     }
 
-//todo move to  parsers
-    template<typename Struct>
-    concept continuousStruct=requires(Struct s)
+
+
+//todo printer for vectors,
+//todo printer for matrix
+//todo find and get solutions form
+
+    template<typename PrintType>
+    int floatPrinter(std::ostream&out,const PrintType& printee,int N=std::numeric_limits<FloatType>::digits10-1);
+
+
+
+    template<typename Container>
+    requires commonDeclarations::isOneDimensionalContinuous<Container> && std::is_floating_point_v<typename Container::value_type>
+    int printSolutionFormat1(std::ostream &out, const Container &solution)
     {
-        requires std::random_access_iterator<decltype(std::begin(s))>;
-    };
+        out << "Решение системы диполей\n Ai(x\\ny)\tBi(x\\ny)\tCi(x\\ny)\n";
+        int N_ = solution.size() / 4.0;
 
-    template<continuousStruct Struct>
-    Struct parseCont(Json::Value &val) {
-        int size = val["size"].asUInt();
+        for (int i = 0; i < N_; i++) {
+            auto cx = sqrt(solution[2 * i] * solution[2 * i] +
+                           solution[2 * i + 2 * N_] * solution[2 * i + 2 * N_]);
+            auto cy = sqrt(solution[2 * i + 1] * solution[2 * i + 1] +
+                           solution[2 * i + 1 + 2 * N_] * solution[2 * i + 1 + 2 * N_]);
+            IosStatePreserve state(out);
+            out << std::scientific;
 
-        std::valarray<double> res(size);
+            out << solution[2 * i] << "\t"
+                << solution[2 * i + 2 * N_] << "\t"
+                << cx << "\n";
+            out << solution[2 * i + 1] << "\t"
+                << solution[2 * i + 1 + 2 * N_] << "\t"
+                << cy << "\n";
 
-        for (int i = 0; i < size; ++i) {
-            res[i] = val["data"][i].asDouble();
         }
-        return res;
+        return 0;
     }
+
+
+
+
+
+
+    template<class T>
+    void printSolution(std::ostream &out, std::vector<T> &solution_, Eigen::IOFormat format = Eigen::IOFormat()) {
+        Eigen::Map<Eigen::Vector<T, Eigen::Dynamic>> map(solution_.data(), solution_.size());//todo copy impl(as template
+        out << "Вектор решения\n" << map.format(format) << "\n";
+    }
+//todo make this one support all structs
+    template<class T>
+    void printSolution(std::ostream &out, Eigen::Vector<T, Eigen::Dynamic> &solution_,
+                       Eigen::IOFormat format = Eigen::IOFormat()) {
+        out << "Вектор решения\n" << solution_.format(format) << "\n";
+    }
+
+
+    template<typename Collection>
+    void printCoordinates2(std::ostream &out, const Collection &xi) {
+        out << "Координаты диполей\n";
+
+        if constexpr (not commonDeclarations::HasBracketsNested<Collection>) {
+            auto N = xi.size() / 2;
+            for (int i = 0; i < N; ++i) {
+                out << xi[i] << '\t' << xi[i + N] << "\n";
+            }
+        } else {
+            auto N = xi[0].size();
+            for (int i = 0; i < N; ++i) {
+                out << xi[0][i] << '\t' << xi[1][i] << "\n";
+            }
+        }
+    }
+
+
+
+
+
 
 
 }
